* partner/dd-staging-polaris: (100 commits)
  Squashed 'drivers/staging/qcacld-3.0/' changes from 561c8433b163..9bf55abaa446
  Squashed 'drivers/staging/qca-wifi-host-cmn/' changes from ef4d961c7452..f48662bbc5d5
  clocksource: arch_timer: make virtual counter access configurable
  arm64: issue isb when trapping CNTVCT_EL0 access
  BACKPORT: arm64: Add CNTFRQ_EL0 trap handler
  BACKPORT: arm64: Add CNTVCT_EL0 trap handler
  qseecom: resend continue blocked listener request
  msm: ipa: Remove unwanted gsi debug register reads
  msm: ipa: Fix the handling of default IPA header
  msm: adsprpc: Use unsigned integer for length values
  arm64: SW PAN: Update saved ttbr0 value on enter_lazy_tlb
  arm64: SW PAN: Point saved ttbr0 at the zero page when switching to init_mm
  msm: jpegdma: Fix for ASAN issues in jpegdma module
  msm: camera: Prevent buffer overread in write_logsync.
  oc: qcom: rpm-smd-debug: Fix potential memory leaks
  ion: ensure CMO target is valid
  power: qpnp-fg-gen3: Use disable_irq_nosync for disabling interrupts
  msm: Array bounds check for buffer index
  msm: ipa3: fix holb sequence for SSR
  ASoC: mbhc: Avoid multiple insertion/removals for aux cable
  ...

Bug: 71899668

+++ drivers/staging/qcacld-3.0/core/dp/ol/inc/ol_txrx_htt_api.h
@@ -419,7 +419,7 @@ void ol_rx_frag_indication_handler(ol_txrx_pdev_handle pdev,
  */
 void
 ol_rx_offload_deliver_ind_handler(ol_txrx_pdev_handle pdev,
-				  qdf_nbuf_t msg, int msdu_cnt);
+				  qdf_nbuf_t msg, uint16_t msdu_cnt);
 
 /**
  * @brief Process a peer map message sent by the target.
@@ -622,8 +622,8 @@ void
 ol_rx_pn_ind_handler(ol_txrx_pdev_handle pdev,
 		     uint16_t peer_id,
 		     uint8_t tid,
-		     int seq_num_start,
-		     int seq_num_end, uint8_t pn_ie_cnt, uint8_t *pn_ie);
+		     uint16_t seq_num_start,
+		     uint16_t seq_num_end, uint8_t pn_ie_cnt, uint8_t *pn_ie);
 
 /**
  * @brief Process a stats message sent by the target.
